requirements.txt
Flask>=3.0,<4.0
Flask-Login>=0.6,<1.0
mysql-connector-python>=9.0,<10.0
Werkzeug>=3.0,<4.0
numpy>=1.26,<3.0
pandas>=2.0,<3.0
scipy>=1.11,<2.0
matplotlib>=3.8,<4.0
bokeh>=3.4,<4.0
panel>=1.4,<2.0
flopy>=3.8,<4.0
gunicorn>=23.0,<24.0
python-dotenv>=1.0,<2.0
reportlab>=4.0,<5.0
